Automatically backed up by Learn

+++ index.js
@@ -26,7 +26,7 @@ function makeInt(a) {
  if  (a.toString() === a) {
   return a;
  }
- if (a.parseInt('0x2328')) {
+ if (a.parseFloat('0x2328')) {
    return 0;
  }
 }
